pyproject.toml
[project]
name = "agro-reich"
version = "0.1.0"
description = ""
authors = [
    {name = "DimonBor",email = "dimonbor200321@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "keyboard (>=0.13.5,<0.14.0)",
    "pyautogui (>=0.9.54,<0.10.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "networkx (>=3.4.2,<4.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "xlib (>=0.21,<0.22)",
    "mss (>=10.0.0,<11.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
]

[tool.poetry]
packages = [{include = "agro_reich", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
